test(integration): Created Common test class for ALL integration tests
ref: #1151 #152 nofusscomputing/centurion_erp_ui#201 nofusscomputing/kubernetes#1
